refactor(shared): split multi-type files into one type per file
Extract co-located types from their parent files so each file declares
exactly one top-level type:

- MealConfirmedIngredient → own file under Shared.Events/CrossModule
- ShoppingListProjectedDetail → own file under Shopping.Application/Interfaces
- FakeMealPlanEventStore, FakeMealPlanReadModelRepository → own files in MealPlan.Application.Tests
- TestCommand, PlainCommand, TestQuery, TestMeterFactory → own files in Shared.Tests/CQRS

Also moves Yumney.Shared.Events.Wolverine into the /src/Shared/ solution
folder, and drops a stray using + tightens an endpoint variable in Shopping.
