+++ src/Yumney.Shared.Events/CrossModule/MealConfirmedIngredient.cs
@@ -1,0 +1,3 @@
+namespace SmartSolutionsLab.Yumney.Shared.Events.CrossModule;
+
+public sealed record MealConfirmedIngredient(string Name, decimal Quantity, string? Unit);
