+++ src/Yumney.Shopping.Application/Interfaces/ShoppingListProjectedDetail.cs
@@ -1,0 +1,5 @@
+using SmartSolutionsLab.Yumney.Shopping.Application.DTOs;
+
+namespace SmartSolutionsLab.Yumney.Shopping.Application.Interfaces;
+
+public sealed record ShoppingListProjectedDetail(string OwnerId, ShoppingListDetailDto Dto);
