+++ tests/Yumney.Shared.Tests/CQRS/TestCommand.cs
@@ -1,0 +1,6 @@
+using SmartSolutionsLab.Yumney.Shared.Common;
+using SmartSolutionsLab.Yumney.Shared.CQRS;
+
+namespace SmartSolutionsLab.Yumney.Shared.Tests.CQRS;
+
+public sealed record TestCommand : ICommand<Result<string>>;
