+++ tests/Yumney.Shared.Tests/CQRS/TestMeterFactory.cs
@@ -1,0 +1,12 @@
+using System.Diagnostics.Metrics;
+
+namespace SmartSolutionsLab.Yumney.Shared.Tests.CQRS;
+
+internal sealed class TestMeterFactory : IMeterFactory
+{
+	public Meter Create(MeterOptions options) => new(options);
+
+	public void Dispose()
+	{
+	}
+}
