+++ tests/Yumney.Shared.Tests/CQRS/TestQuery.cs
@@ -1,0 +1,6 @@
+using SmartSolutionsLab.Yumney.Shared.Common;
+using SmartSolutionsLab.Yumney.Shared.CQRS;
+
+namespace SmartSolutionsLab.Yumney.Shared.Tests.CQRS;
+
+public sealed record TestQuery : IQuery<Result<int>>;
